remove old logic funcs
