+++ common/functions/src/scalars/function2_factory.rs
@@ -26,7 +26,7 @@ use super::function_factory::FunctionFeatures;
 use super::ComparisonFunction;
 use super::ConditionalFunction;
 use super::HashesFunction;
-use super::LogicFunction2;
+use super::LogicFunction;
 use super::StringFunction;
 use super::TupleClassFunction;
 use super::UdfFunction;
@@ -69,7 +69,7 @@ static FUNCTION2_FACTORY: Lazy<Arc<Function2Factory>> = Lazy::new(|| {
     StringFunction::register2(&mut function_factory);
     HashesFunction::register2(&mut function_factory);
     ConditionalFunction::register(&mut function_factory);
-    LogicFunction2::register(&mut function_factory);
+    LogicFunction::register(&mut function_factory);
 
     Arc::new(function_factory)
 });
