+++ common/functions/src/scalars/function_factory.rs
@@ -151,7 +151,6 @@ static FUNCTION_FACTORY: Lazy<Arc<FunctionFactory>> = Lazy::new(|| {
     let mut function_factory = FunctionFactory::create();
 
     ArithmeticFunction::register(&mut function_factory);
-    LogicFunction::register(&mut function_factory);
     NullableFunction::register(&mut function_factory);
     StringFunction::register(&mut function_factory);
     UdfFunction::register(&mut function_factory);
