+++ common/functions/tests/it/scalars/logics.rs
@@ -12,53 +12,150 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common_datavalues::prelude::*;
+use std::sync::Arc;
+
+use common_datavalues2::prelude::*;
 use common_exception::Result;
-use common_functions::scalars::*;
+use common_functions::scalars::LogicAndFunction;
+use common_functions::scalars::LogicNotFunction;
+use common_functions::scalars::LogicOrFunction;
+use common_functions::scalars::LogicXorFunction;
 
-use crate::scalars::scalar_function_test::test_scalar_functions;
-use crate::scalars::scalar_function_test::ScalarFunctionTest;
+use crate::scalars::scalar_function2_test::test_scalar_functions2;
+use crate::scalars::scalar_function2_test::ScalarFunction2Test;
 
-// TODO: need test null and logic functions
+#[test]
+fn test_logic_not_function() -> Result<()> {
+    let tests = vec![
+        ScalarFunction2Test {
+            name: "not",
+            columns: vec![Series::from_data(vec![true, false]).into()],
+            expect: Series::from_data(vec![false, true]).into(),
+            error: "",
+        },
+        ScalarFunction2Test {
+            name: "not-null",
+            columns: vec![Series::from_data(vec![None, Some(true), Some(false)]).into()],
+            expect: Series::from_data(vec![None, Some(false), Some(true)]).into(),
+            error: "",
+        },
+        ScalarFunction2Test {
+            name: "not-null",
+            columns: vec![Arc::new(NullColumn::new(4))],
+            expect: Arc::new(NullColumn::new(4)),
+            error: "",
+        },
+    ];
+    test_scalar_functions2(LogicNotFunction::try_create("not")?, &tests)
+}
 
 #[test]
 fn test_logic_and_function() -> Result<()> {
-    let tests = vec![ScalarFunctionTest {
-        name: "and-passed",
-        nullable: false,
-        columns: vec![
-            Series::new(vec![true, true, true, false]).into(),
-            Series::new(vec![true, false, true, true]).into(),
-        ],
-        expect: Series::new(vec![true, false, true, false]).into(),
-        error: "",
-    }];
-    test_scalar_functions(LogicAndFunction::try_create_func("")?, &tests)
+    let tests = vec![
+        ScalarFunction2Test {
+            name: "and",
+            columns: vec![
+                Series::from_data(vec![true, true, true, false]),
+                Series::from_data(vec![true, false, true, true]),
+            ],
+            expect: Series::from_data(vec![true, false, true, false]),
+            error: "",
+        },
+        ScalarFunction2Test {
+            name: "and-null",
+            columns: vec![
+                Series::from_data(vec![None, Some(true), Some(true), Some(false)]),
+                Series::from_data(vec![true, false, true, true]),
+            ],
+            expect: Series::from_data(vec![None, Some(false), Some(true), Some(false)]),
+            error: "",
+        },
+        ScalarFunction2Test {
+            name: "and-null",
+            columns: vec![
+                Series::from_data(vec![None, Some(true), Some(true), Some(false)]),
+                Arc::new(NullColumn::new(4)),
+            ],
+            expect: Arc::new(NullColumn::new(4)),
+            error: "",
+        },
+    ];
+    test_scalar_functions2(LogicAndFunction::try_create("and")?, &tests)
 }
 
 #[test]
 fn test_logic_or_function() -> Result<()> {
-    let tests = vec![ScalarFunctionTest {
-        name: "or-passed",
-        nullable: false,
-        columns: vec![
-            Series::new(vec![true, true, true, false]).into(),
-            Series::new(vec![true, false, true, true]).into(),
-        ],
-        expect: Series::new(vec![true, true, true, true]).into(),
-        error: "",
-    }];
-    test_scalar_functions(LogicOrFunction::try_create_func("")?, &tests)
+    let tests = vec![
+        ScalarFunction2Test {
+            name: "or",
+            columns: vec![
+                Series::from_data(vec![true, true, true, false]),
+                Series::from_data(vec![true, false, true, false]),
+            ],
+            expect: Series::from_data(vec![true, true, true, false]),
+            error: "",
+        },
+        ScalarFunction2Test {
+            name: "or-null",
+            columns: vec![
+                Series::from_data(vec![None, None, None, Some(false)]),
+                Series::from_data(vec![Some(true), Some(false), None, Some(true)]),
+            ],
+            expect: Series::from_data(vec![Some(true), None, None, Some(true)]),
+            error: "",
+        },
+        ScalarFunction2Test {
+            name: "or-null",
+            columns: vec![
+                Series::from_data(vec![None, None, None, Some(false)]),
+                Series::from_data(vec![Some(true), Some(false), None, Some(true)]),
+            ],
+            expect: Series::from_data(vec![Some(true), None, None, Some(true)]),
+            error: "",
+        },
+        ScalarFunction2Test {
+            name: "or-null",
+            columns: vec![
+                Arc::new(NullColumn::new(4)),
+                Series::from_data(vec![Some(true), None, None, Some(false)]),
+            ],
+            expect: Series::from_data(vec![Some(true), None, None, None]),
+            error: "",
+        },
+    ];
+    test_scalar_functions2(LogicOrFunction::try_create("or")?, &tests)
 }
 
 #[test]
-fn test_logic_not_function() -> Result<()> {
-    let tests = vec![ScalarFunctionTest {
-        name: "not-passed",
-        nullable: false,
-        columns: vec![Series::new(vec![true, false]).into()],
-        expect: Series::new(vec![false, true]).into(),
-        error: "",
-    }];
-    test_scalar_functions(LogicNotFunction::try_create_func("")?, &tests)
+fn test_logic_xor_function() -> Result<()> {
+    let tests = vec![
+        ScalarFunction2Test {
+            name: "xor",
+            columns: vec![
+                Series::from_data(vec![true, true, false, false]),
+                Series::from_data(vec![true, false, true, false]),
+            ],
+            expect: Series::from_data(vec![false, true, true, false]),
+            error: "",
+        },
+        ScalarFunction2Test {
+            name: "xor-null",
+            columns: vec![
+                Series::from_data(vec![None, Some(true), Some(false), Some(false)]),
+                Series::from_data(vec![Some(true), None, Some(true), Some(false)]),
+            ],
+            expect: Series::from_data(vec![None, None, Some(true), Some(false)]),
+            error: "",
+        },
+        ScalarFunction2Test {
+            name: "xor-null",
+            columns: vec![
+                Series::from_data(vec![None, Some(true), Some(false), Some(false)]),
+                Arc::new(NullColumn::new(4)),
+            ],
+            expect: Arc::new(NullColumn::new(4)),
+            error: "",
+        },
+    ];
+    test_scalar_functions2(LogicXorFunction::try_create("or")?, &tests)
 }
