+++ common/functions/tests/it/scalars/mod.rs
@@ -20,7 +20,6 @@ mod expressions;
 mod hashes;
 mod helpers;
 mod logics;
-mod logics2;
 mod maths;
 mod nullables;
 mod others;
